# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Absensi Karyawan GPS & Barcode API",
        "description": "API Documentation untuk Sistem Absensi Karyawan dengan GPS dan Barcode Scanner",
        "contact": {
            "name": "Admin Absensi",
            "email": "admin@absensi.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/admin/import-export/users": {
            "get": {
                "tags": [
                    "Admin - Import/Export"
                ],
                "summary": "Import/Export users",
                "description": "Menampilkan halaman untuk import dan export data users",
                "operationId": "7e1a617be2026f64c78ade094375e5b3",
                "responses": {
                    "200": {
                        "description": "Success - Returns HTML view"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/import-export/attendances": {
            "get": {
                "tags": [
                    "Admin - Import/Export"
                ],
                "summary": "Import/Export attendances",
                "description": "Menampilkan halaman untuk import dan export data absensi",
                "operationId": "1c17e6312671cbe7fbbca2ca77a0e248",
                "responses": {
                    "200": {
                        "description": "Success - Returns HTML view"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/masterdata/division": {
            "get": {
                "tags": [
                    "Admin - Master Data"
                ],
                "summary": "Master data divisi",
                "description": "Menampilkan halaman manajemen master data divisi",
                "operationId": "1f8439d9791cd5757e6146a1c25fbe27",
                "responses": {
                    "200": {
                        "description": "Success - Returns HTML view"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/masterdata/job-title": {
            "get": {
                "tags": [
                    "Admin - Master Data"
                ],
                "summary": "Master data jabatan",
                "description": "Menampilkan halaman manajemen master data jabatan",
                "operationId": "a8e6490945943da56aaab742dcf312f8",
                "responses": {
                    "200": {
                        "description": "Success - Returns HTML view"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/masterdata/education": {
            "get": {
                "tags": [
                    "Admin - Master Data"
                ],
                "summary": "Master data pendidikan",
                "description": "Menampilkan halaman manajemen master data pendidikan",
                "operationId": "ddbc79c5988758cf7fd389de83b08836",
                "responses": {
                    "200": {
                        "description": "Success - Returns HTML view"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/masterdata/shift": {
            "get": {
                "tags": [
                    "Admin - Master Data"
                ],
                "summary": "Master data shift",
                "description": "Menampilkan halaman manajemen master data shift kerja",
                "operationId": "99fbcb4cbbad60dc2b26e5b2ad8add67",
                "responses": {
                    "200": {
                        "description": "Success - Returns HTML view"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/masterdata/admin": {
            "get": {
                "tags": [
                    "Admin - Master Data"
                ],
                "summary": "Master data admin",
                "description": "Menampilkan halaman manajemen master data admin",
                "operationId": "7813d09f775cc0198291186fd3793e41",
                "responses": {
                    "200": {
                        "description": "Success - Returns HTML view"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/home": {
            "get": {
                "tags": [
                    "User - Dashboard"
                ],
                "summary": "Dashboard user",
                "description": "Menampilkan dashboard utama untuk karyawan",
                "operationId": "57db05a8a6f7b699e797fbb12aef816e",
                "responses": {
                    "200": {
                        "description": "Success - Returns HTML view"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/apply-leave": {
            "get": {
                "tags": [
                    "User - Attendance"
                ],
                "summary": "Form pengajuan cuti/izin",
                "description": "Menampilkan form untuk pengajuan cuti atau izin karyawan",
                "operationId": "45fe71691bdbe1b7a03dc353c19298b6",
                "responses": {
                    "200": {
                        "description": "Success - Returns HTML view"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Account not verified"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User - Attendance"
                ],
                "summary": "Submit pengajuan cuti/izin",
                "description": "Menyimpan pengajuan cuti atau izin karyawan",
                "operationId": "8e5f5dd1b6741ae06ed100bcfc96e221",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "status",
                                    "note",
                                    "from"
                                ],
                                "properties": {
                                    "status": {
                                        "description": "Jenis izin",
                                        "type": "string",
                                        "enum": [
                                            "excused",
                                            "sick"
                                        ]
                                    },
                                    "note": {
                                        "description": "Keterangan",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "from": {
                                        "description": "Tanggal mulai",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "to": {
                                        "description": "Tanggal selesai (opsional)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "attachment": {
                                        "description": "File lampiran (max 3MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "lat": {
                                        "description": "Latitude lokasi",
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "lng": {
                                        "description": "Longitude lokasi",
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect to home dengan flash message"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Account not verified"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/attendance-history": {
            "get": {
                "tags": [
                    "User - Attendance"
                ],
                "summary": "Riwayat absensi",
                "description": "Menampilkan riwayat absensi karyawan",
                "operationId": "40524c07bdc65c1f2b793dbfa263535c",
                "responses": {
                    "200": {
                        "description": "Success - Returns HTML view"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}